refactor: update imports, add userId to tests, transaction repo, service and DTOs
