+++ src/internal/dto/transaction/create-transaction.dto.ts
@@ -1,4 +1,5 @@
 export interface CreateTransactionDto {
   amount: number;
   type: "IN" | "OUT";
+  userId: string;
 }
