+++ src/internal/service/transaction/create-transaction.ts
@@ -1,6 +1,6 @@
-import { TransactionRepository } from "@/factory/repository/transaction";
 import { CreateTransactionDto } from "@/internal/dto/transaction";
 import InvalidDataError from "@/internal/error/invalid-data";
+import { TransactionRepository } from "@/internal/interface/repository/transaction";
 import { validateTransaction } from "@/internal/validation/transaction";
 
 export default async function createTransaction(
