+++ src/internal/service/transaction/delete-transaction.ts
@@ -1,5 +1,5 @@
-import { TransactionRepository } from "@/factory/repository/transaction";
 import InvalidDataError from "@/internal/error/invalid-data";
+import { TransactionRepository } from "@/internal/interface/repository/transaction";
 import { validateTransaction } from "@/internal/validation/transaction";
 
 export default async function deleteTransaction(
