+++ src/internal/validation/transaction/__tests__/validate-create-data.spec.ts
@@ -7,6 +7,7 @@ describe("transaction create data", () => {
     const invalidData = {
       amount: "invalid",
       type: "INVALID",
+      userId: "invalid",
     };
     // @ts-ignore
     const isValid = createData(invalidData);
@@ -17,6 +18,7 @@ describe("transaction create data", () => {
     const validData: CreateTransactionDto = {
       amount: 100,
       type: "IN",
+      userId: "c6a26ac0-4fc1-4fbb-8de4-835b21e59e00",
     };
 
     const isValid = createData(validData);
