+++ src/repository/transaction/create-transaction.ts
@@ -4,11 +4,13 @@ import { CreateTransactionDto } from "@/internal/dto/transaction";
 export default async function createTransaction({
   amount,
   type,
+  userId,
 }: CreateTransactionDto) {
   const transaction = await prismaClient.transaction.create({
     data: {
       amount,
       type,
+      userId,
     },
   });
   return transaction;
